[system] network diagnostics for Gen 3 platforms
