+++ system/src/system_network_internal.cpp
@@ -65,21 +65,3 @@ void resetNetworkInterfaces() {
 }
 
 } // namespace particle
-
-/* FIXME: there should be a define that tells whether there is NetworkManager available
- * or not */
-#if !HAL_PLATFORM_IFAPI
-
-namespace {
-
-using namespace particle;
-
-NetworkDiagnostics g_networkDiagnostics;
-
-} // namespace
-
-particle::NetworkDiagnostics* particle::NetworkDiagnostics::instance() {
-    return &g_networkDiagnostics;
-}
-
-#endif /* !HAL_PLATFORM_IFAPI */
