+++ system/src/system_network_internal.h
@@ -90,7 +90,7 @@ void resetNetworkInterfaces();
 #include "rgbled.h"
 #include "spark_wiring_led.h"
 #include "spark_wiring_ticks.h"
-#include "spark_wiring_diagnostics.h"
+#include "system_network_diagnostics.h"
 #include "system_event.h"
 #include "system_cloud_internal.h"
 #include "system_network.h"
@@ -130,73 +130,6 @@ private:
 #else // !defined(DEBUG_NETWORK_STATE)
 #define LOG_NETWORK_STATE()
 #endif
-
-namespace particle {
-
-class NetworkDiagnostics {
-public:
-    // Note: Use odd numbers to encode transitional states
-    enum Status {
-        TURNED_OFF = 0,
-        TURNING_ON = 1,
-        DISCONNECTED = 2,
-        CONNECTING = 3,
-        CONNECTED = 4,
-        DISCONNECTING = 5,
-        TURNING_OFF = 7
-    };
-
-    NetworkDiagnostics() :
-            status_(DIAG_ID_NETWORK_CONNECTION_STATUS, DIAG_NAME_NETWORK_CONNECTION_STATUS, TURNED_OFF),
-            disconnReason_(DIAG_ID_NETWORK_DISCONNECTION_REASON, DIAG_NAME_NETWORK_DISCONNECTION_REASON, NETWORK_DISCONNECT_REASON_NONE),
-            disconnCount_(DIAG_ID_NETWORK_DISCONNECTS, DIAG_NAME_NETWORK_DISCONNECTS),
-            connCount_(DIAG_ID_NETWORK_CONNECTION_ATTEMPTS, DIAG_NAME_NETWORK_CONNECTION_ATTEMPTS),
-            lastError_(DIAG_ID_NETWORK_CONNECTION_ERROR_CODE, DIAG_NAME_NETWORK_CONNECTION_ERROR_CODE) {
-    }
-
-    NetworkDiagnostics& status(Status status) {
-        status_ = status;
-        return *this;
-    }
-
-    NetworkDiagnostics& connectionAttempt() {
-        ++connCount_;
-        return *this;
-    }
-
-    NetworkDiagnostics& resetConnectionAttempts() {
-        connCount_ = 0;
-        return *this;
-    }
-
-    NetworkDiagnostics& disconnectionReason(network_disconnect_reason reason) {
-        disconnReason_ = reason;
-        return *this;
-    }
-
-    NetworkDiagnostics& disconnectedUnexpectedly() {
-        ++disconnCount_;
-        return *this;
-    }
-
-    NetworkDiagnostics& lastError(int error) {
-        lastError_ = error;
-        return *this;
-    }
-
-    static NetworkDiagnostics* instance();
-
-private:
-    // Some of the diagnostic data sources use the synchronization since they can be updated from
-    // the networking service thread
-    AtomicEnumDiagnosticData<Status> status_;
-    AtomicEnumDiagnosticData<network_disconnect_reason> disconnReason_;
-    AtomicUnsignedIntegerDiagnosticData disconnCount_;
-    SimpleUnsignedIntegerDiagnosticData connCount_;
-    SimpleIntegerDiagnosticData lastError_;
-};
-
-} // namespace particle
 
 /**
  * Internal network interface class to provide polymorphic behavior for each
